requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
watchdog==4.0.0
google-api-python-client==2.162.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
gspread==5.10.0